requirements.txt
Django

djangorestframework
djangorestframework-gis

django-filter
django-extensions

psycopg2-binary

tqdm
requests
PyYAML
python-dateutil
